refactor(vue-renderer): split renderer into ssr, spa and modern
