+++ packages/server/src/middleware/modern.js
@@ -1,5 +1,3 @@
-import chalk from 'chalk'
-import consola from 'consola'
 import UAParser from 'ua-parser-js'
 import semver from 'semver'
 
@@ -23,22 +21,6 @@ const isModernBrowser = (ua) => {
   return Boolean(modernBrowsers[browser.name] && semver.gte(browserVersion, modernBrowsers[browser.name]))
 }
 
-const distinctModernModeOptions = [false, 'client', 'server']
-
-const detectModernBuild = ({ options, resources }) => {
-  if (distinctModernModeOptions.includes(options.modern)) {
-    return
-  }
-
-  if (!resources.modernManifest) {
-    options.modern = false
-    return
-  }
-
-  options.modern = options.render.ssr ? 'server' : 'client'
-  consola.info(`Modern bundles are detected. Modern mode (${chalk.green.bold(options.modern)}) is enabled now.`)
-}
-
 const detectModernBrowser = ({ socket = {}, headers }) => {
   if (socket.isModernBrowser === undefined) {
     const ua = headers && headers['user-agent']
@@ -49,14 +31,9 @@ const detectModernBrowser = ({ socket = {}, headers }) => {
 }
 
 export default ({ context }) => {
-  let detected = false
   return (req, res, next) => {
-    if (!detected) {
-      detectModernBuild(context)
-      detected = true
-    }
     if (context.options.modern !== false) {
-      req.modernMode = detectModernBrowser(req)
+      req._modern = detectModernBrowser(req)
     }
     next()
   }
