+++ packages/server/src/middleware/nuxt.js
@@ -25,7 +25,7 @@ export default ({ options, nuxt, renderRoute, resources }) => async function nux
       cspScriptSrcHashes,
       error,
       redirected,
-      getPreloadFiles
+      preloadFiles
     } = result
 
     if (redirected) {
@@ -52,8 +52,6 @@ export default ({ options, nuxt, renderRoute, resources }) => async function nux
     if (!error && options.render.http2.push) {
       // Parse resourceHints to extract HTTP.2 prefetch/push headers
       // https://w3c.github.io/preload/#server-push-http-2
-      const preloadFiles = getPreloadFiles()
-
       const { shouldPush, pushAssets } = options.render.http2
       const { publicPath } = resources.clientManifest
